from double to float

+++ block.cl
@@ -10,7 +10,7 @@
 #define BS(i,j) bs[j + i*BLOCK_SIZE]
 
 __kernel __attribute__ ((reqd_work_group_size(BLOCK_SIZE, BLOCK_SIZE, 1)))
-void block(__global double* a, __global double* b, __global double* output, __local double* as, __local double* bs)
+void block(__global float* a, __global float* b, __global float* output, __local float* as, __local float* bs)
 {
     // Index of the tile/work-group
     int block_x = get_group_id(0);
@@ -21,7 +21,7 @@ void block(__global double* a, __global double* b, __global double* output, __lo
     int thread_y = get_local_id(1);
 
     int rank = get_global_size(0);
-    double running = 0.0f;
+    float running = 0.0f;
 
    // Starting index for a and b matrices i.e. first sub matrices
     int a_index = rank * BLOCK_SIZE * block_y;
